# Perf: Use DateTime internally