--- a/src/LetsTrace/Util/IClock.cs
+++ b/src/LetsTrace/Util/IClock.cs
@@ -4,6 +4,6 @@ namespace LetsTrace.Util
 {
     public interface IClock
     {
-        DateTimeOffset CurrentTime();
+        DateTime UtcNow();
     }
 }