… src/LetsTrace/Util/DateTimeOffsetExtensionMethods.cs
… test/LetsTrace.Tests/SpanTests.cs