The main reason why we used `DateTimeOffset` for the API in OpenTracing is that it is more user-friendly. If we had exposed `DateTime` we wouldn't have had a way to ensure that the user always passes objects with `DateTimeKind.Utc`. See https://github.com/opentracing/opentracing-csharp/issues/42 for more details.

A tracer however does not have to use `DateTimeOffset` internally as it doesn't need the actual time zone difference - it only needs UTC values. Using `DateTime` with `DateTimeKind.Utc` instead therefore gives better performance!

I'm not doing any validations of the `DateTimeKind` on the `DateTime` objects as all of it is controlled within this code (No one can pass a DateTime with e.g. DateTimeKind.Local or Unspecified). I renamed the properties to make this more clear though.)

#### Before:
``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
```
|                         Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------- |-----------:|---------:|---------:|-------:|----------:|
|              NoTimestamp_Start |   919.8 ns | 19.83 ns | 34.20 ns | 0.5322 |     840 B |
|       NoTimestamp_Start_Finish | 1,002.1 ns | 19.99 ns | 22.22 ns | 0.5627 |     888 B |
|   NoTimestamp_Start_Log_Finish | 1,309.3 ns | 23.42 ns | 21.91 ns | 0.7629 |    1200 B |
|            WithTimestamp_Start | 1,581.7 ns | 23.51 ns | 21.99 ns | 0.5322 |     840 B |
|     WithTimestamp_Start_Finish | 2,315.7 ns | 32.00 ns | 26.72 ns | 0.5608 |     888 B |
| WithTimestamp_Start_Log_Finish | 3,334.2 ns | 64.99 ns | 66.74 ns | 0.7629 |    1200 B |

#### After:

``` ini
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.309)
Intel Core i5-4300U CPU 1.90GHz (Haswell), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2435766 Hz, Resolution=410.5485 ns, Timer=TSC
.NET Core SDK=2.1.300-preview1-008174
  [Host]     : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.6 (CoreCLR 4.6.26212.01, CoreFX 4.6.26212.01), 64bit RyuJIT
```
|                         Method |       Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------- |-----------:|----------:|----------:|-------:|----------:|
|              NoTimestamp_Start |   872.5 ns | 16.053 ns | 15.016 ns | 0.5178 |     816 B |
|       NoTimestamp_Start_Finish |   936.4 ns |  9.528 ns |  7.956 ns | 0.5484 |     864 B |
|   NoTimestamp_Start_Log_Finish | 1,067.6 ns | 18.927 ns | 17.704 ns | 0.6962 |    1096 B |
|            WithTimestamp_Start | 1,566.8 ns | 28.915 ns | 27.047 ns | 0.5169 |     816 B |
|     WithTimestamp_Start_Finish | 2,345.7 ns | 43.614 ns | 40.796 ns | 0.5455 |     864 B |
| WithTimestamp_Start_Log_Finish | 3,177.3 ns | 43.618 ns | 38.666 ns | 0.6943 |    1096 B |